test-store: Make tools to store test blocks available
Move the tools used to set up fake blocks and test against them from the
chain_head tests to the test store.
